backend/requirements.txt
fastapi>=0.109.0
sqlmodel>=0.0.14
python-jose[cryptography]>=3.3.0
psycopg2-binary>=2.9.9
uvicorn>=0.25.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pytest>=7.4.0
httpx>=0.25.0
pytest-asyncio>=0.21.0
